requirements.txt
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
Flask==3.1.3
google-auth==2.49.1
google-genai==1.68.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
pillow==12.1.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyperclip==1.11.0
python-dotenv==1.2.2
requests==2.32.5
sniffio==1.3.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.6
